requirements.txt
setuptools>=80.0.0
gldb==2.1.1
appdirs~=1.4.4
ontolutils>=0.25.1,<1.0.0
click~=8.1.7
rdflib~=7.1.1
requests~=2.32.4
h5rdmtoolbox>=2.5.3,<=3.0.0
ssnolib>=2.2.0.2,<=3.0.0.0
tqdm~=4.67.1
rdflib-endpoint[cli]
python-dotenv
pyshacl>=0.30.1
SPARQLWrapper>=2.0.0,<3.0.0